refactor(runtime): extract network-fault tests into dedicated module (#3212)
