+++ crates/kamn-core/tests/runtime_network_docs.rs
@@ -8,6 +8,7 @@ fn doc_contains_runtime_network_scope_and_models() {
     assert!(DOC.contains("crates/kamn-core/src/runtime_transport_coordination.rs"));
     assert!(DOC.contains("crates/kamn-core/src/runtime_snapshot_store.rs"));
     assert!(DOC.contains("crates/kamn-core/src/runtime_tests_snapshot_store.rs"));
+    assert!(DOC.contains("crates/kamn-core/src/runtime_tests_network_fault.rs"));
     assert!(DOC.contains("crates/kamn-core/src/runtime_recovery_guard.rs"));
     assert!(DOC.contains("## Node CLI Recovery-Check Mapping"));
     assert!(DOC.contains("## Node CLI Daemon Lifecycle Mapping"));
